Rename selector to locator
[ghstack-poisoned]
