+++ package.json
@@ -20,6 +20,7 @@
   "dependencies": {},
   "devDependencies": {
     "@types/node": "^16.0.0",
+    "@types/jest": "^29.0.0",
     "@typescript-eslint/eslint-plugin": "^5.30.0",
     "@typescript-eslint/parser": "^5.30.0",
     "eslint": "^8.32.0",
