+++ packages/core/src/errors/TooManyMatchingElementError.ts
@@ -1,9 +1,9 @@
-import { PartSelectorType } from '../selectors/PartSelectorType';
+import { PartLocatorType } from '../locators/PartLocatorType';
 
 export const TooManyMatchingElementErrorId = 'TooManyMatchingElementError';
 
 export class TooManyMatchingElementError extends Error {
-  constructor(public readonly query: PartSelectorType) {
+  constructor(public readonly query: PartLocatorType) {
     super('Too many matching element');
     this.name = TooManyMatchingElementErrorId;
   }
