+++ packages/core/src/index.ts
@@ -2,9 +2,12 @@ export { ComponentDriver, IntegrationTestEngine } from './ComponentDriver';
 export { defaultOnFinishUpdate, defaultStep } from './defaultValues';
 export type { IInputDriver } from './driverTypes';
 export { TooManyMatchingElementError, TooManyMatchingElementErrorId } from './errors/TooManyMatchingElementError';
-export { byDataTestId } from './selectors/byDataTestId';
-export type { CssSelector, PartSelectorType } from './selectors/PartSelectorType';
-export { SelectorRelativePosition, SelectorType } from './selectors/PartSelectorType';
+export { byDataTestId } from './locators/byDataTestId';
+export type { CssLocator as CssSelector, PartLocatorType as PartSelectorType } from './locators/PartLocatorType';
+export {
+  LocatorRelativePosition as SelectorRelativePosition,
+  LocatorType as SelectorType,
+} from './locators/PartLocatorType';
 export { SimpleComponentDriver } from './SimpleComponentDriver';
 export type { ScenePart, ScenePartDriver } from './types';
 export type { IComponentDriverOption, ITestEngine, ITestEngineOption, StepFunction } from './types';
