+++ packages/core/src/locators/PartLocatorType.ts
@@ -1,0 +1,17 @@
+export enum LocatorRelativePosition {
+  documentRoot = 'documentRoot',
+  baseElement = 'baseElement',
+}
+
+export enum LocatorType {
+  css = 'css',
+  xpath = 'xpath',
+}
+
+export type CssLocator = {
+  type: LocatorType.css;
+  selector: string;
+  relative?: LocatorRelativePosition;
+};
+
+export type PartLocatorType = string | CssLocator;
