--- packages/core/src/selectors/PartSelectorType.ts
@@ -1,17 +1,0 @@
-export enum SelectorRelativePosition {
-  documentRoot = 'documentRoot',
-  baseElement = 'baseElement',
-}
-
-export enum SelectorType {
-  css = 'css',
-  xpath = 'xpath',
-}
-
-export type CssSelector = {
-  type: SelectorType.css;
-  selector: string;
-  relative?: SelectorRelativePosition;
-};
-
-export type PartSelectorType = string | CssSelector;
