+++ packages/integration-driver-mui-v5/__tests__/components/SelectComponentDriver.test.tsx
@@ -29,7 +29,7 @@ const Component = () => {
 
 const testScenePart = {
   select: {
-    selector: byDataTestId('select'),
+    locator: byDataTestId('select'),
     driver: SelectComponentDriver,
   },
 } satisfies ScenePart;
